refactor(skills): drop agent→skill MCP tool, keep UI + API
- remove draft_skill_from_agent MCP tool, rbac entry, schemas, tests
- revert scoped create_skill (6f03288ea): personal-only again, scope helpers back in REST
- regen MCP docs; trim agent docs
