+++ docs/pages/platform-archestra-mcp-server.md
@@ -3,7 +3,7 @@ title: "Archestra MCP Server"
 category: MCP
 description: "Built-in MCP server providing tools for managing Archestra platform resources"
 order: 5
-lastUpdated: 2026-06-01
+lastUpdated: 2026-06-02
 ---
 <!--
 This file is auto-generated by `pnpm codegen:archestra-mcp-server-docs`.
@@ -51,7 +51,6 @@ This tool takes no arguments.
 |------|-------------|--------------------------|
 | `create_agent` | Create a new agent with the specified name, optional description, labels, prompts, icon emoji, explicit tool assignments, and sub-agent delegations. | `agent:create` |
 | `get_agent` | Get a specific agent by ID or name. | `agent:read` |
-| `draft_skill_from_agent` | Convert an internal agent into a draft SKILL.md manifest for review. | `agent:read` |
 | `list_agents` | List agents with optional filtering by name. | `agent:read` |
 | `edit_agent` | Edit an existing agent. | `agent:update` |
 
@@ -125,30 +124,6 @@ Required RBAC permission: `agent:read`
 | `suggestedPrompts` | `object[]` | Yes | Configured suggested prompts. |
 | `suggestedPrompts[].summaryTitle` | `string` | Yes | The short title shown in the chat UI. |
 | `suggestedPrompts[].prompt` | `string` | Yes | The suggested prompt text. |
-
-#### draft_skill_from_agent
-
-Required RBAC permission: `agent:read`
-
-##### Input
-
-| Parameter | Type | Required | Description |
-|-----------|------|----------|-------------|
-| `id` | `string` | Yes | The ID of the agent to convert into a skill draft. Use list_agents or get_agent to look it up by name. |
-
-##### Output
-
-| Field | Type | Required | Description |
-|-------|------|----------|-------------|
-| `manifest` | `string` | Yes | The complete SKILL.md manifest, as data. Review and edit it, then pass it to create_skill to persist. Keep the metadata block to preserve the link to the origin agent. |
-| `scope` | `"personal" \| "team" \| "org"` | Yes | The source agent's scope. Pass it to create_skill (with teamIds when "team") to give the skill the same visibility; omit there to default to a personal skill. Sharing still requires the matching permission. |
-| `teamIds` | `string[]` | Yes | Teams the agent is scoped to — pass these to create_skill alongside scope "team". Empty for personal/org agents. |
-| `carried` | `object[]` | Yes | Agent fields carried over directly to native skill fields. |
-| `carried[].field` | `string` | Yes |  |
-| `carried[].detail` | `string` | Yes |  |
-| `annotated` | `object[]` | Yes | Agent fields with no skill equivalent — folded into metadata or reported as not carried. |
-| `annotated[].field` | `string` | Yes |  |
-| `annotated[].detail` | `string` | Yes |  |
 
 #### list_agents
 
@@ -1598,7 +1573,7 @@ Required RBAC permission: None (no additional RBAC permission required)
 | `list_skills` | List the Agent Skills available in this organization — one line per skill (name and description). | `skill:read` |
 | `activate_skill` | Load a specialized Agent Skill — a reusable SKILL.md instruction set. | `skill:read` |
 | `read_skill_file` | Read a bundled resource file from a skill. | `skill:read` |
-| `create_skill` | Create a new Agent Skill from a SKILL.md manifest, available via list_skills and as a chat slash-command. | `skill:create` |
+| `create_skill` | Create a new Agent Skill from a SKILL.md manifest. | `skill:create` |
 | `update_skill` | Update an existing Agent Skill from a SKILL.md manifest. | `skill:update` |
 
 #### list_skills
@@ -1644,8 +1619,6 @@ Required RBAC permission: `skill:create`
 | `files[].path` | `string` | Yes | Resource path, e.g. references/API.md or scripts/run.py |
 | `files[].content` | `string` | Yes | Text content of the file |
 | `files[].encoding` | `"utf8" \| "base64"` | No |  |
-| `scope` | `"personal" \| "team" \| "org"` | No | Visibility scope. Defaults to `personal` (only you). `team` shares it with the teams in `teamIds` and requires team-admin permission in each; `org` shares it organization-wide and requires admin. Omit unless the user asked to share the skill. |
-| `teamIds` | `string[]` | No | Teams to share a `team`-scoped skill with. Required (non-empty) when scope is `team`; ignored otherwise. |
 
 
 #### update_skill
