+++ platform/backend/src/routes/skill.ts
@@ -2,6 +2,7 @@ import {
   calculatePaginationMeta,
   createPaginatedResponseSchema,
   PaginationQuerySchema,
+  type ResourceVisibilityScope,
   ResourceVisibilityScopeSchema,
   RouteId,
 } from "@shared";
@@ -12,11 +13,9 @@ import {
   requireAgentModifyPermission,
 } from "@/auth/agent-type-permissions";
 import {
-  assertSkillTeams,
-  authorizeSkillScope,
   getSkillPermissionChecker,
   requireSkillModifyPermission,
-  withTeamFkErrorMapped,
+  type SkillPermissionChecker,
 } from "@/auth/skill-permissions";
 import { withDbTransaction } from "@/database";
 import logger from "@/logging";
@@ -59,6 +58,7 @@ import {
   SkillWithFilesSchema,
   UuidIdSchema,
 } from "@/types";
+import { isForeignKeyConstraintError } from "@/utils/db";
 
 /** A team a skill is assigned to (for `scope = 'team'` skills). */
 const SkillTeamSchema = z.object({ id: z.string(), name: z.string() });
@@ -955,6 +955,92 @@ function sameTeamSet(a: string[], b: string[]): boolean {
   if (a.length !== b.length) return false;
   const setB = new Set(b);
   return a.every((id) => setB.has(id));
+}
+
+/**
+ * Validate a skill's team assignments before persisting. Only meaningful for
+ * `team` scope: such a skill must have at least one team (otherwise it is
+ * invisible to everyone, including its author), and every team must exist
+ * within the organization — a stale/deleted id fails with a clean 400 instead
+ * of an FK violation mid-transaction.
+ */
+async function assertSkillTeams(params: {
+  scope: ResourceVisibilityScope;
+  teamIds: string[];
+  organizationId: string;
+}): Promise<void> {
+  if (params.scope !== "team") return;
+
+  if (params.teamIds.length === 0) {
+    throw new ApiError(
+      400,
+      "A team-scoped skill must be assigned to at least one team",
+    );
+  }
+
+  const teams = await TeamModel.findByIds(params.teamIds);
+  const validIds = new Set(
+    teams
+      .filter((team) => team.organizationId === params.organizationId)
+      .map((team) => team.id),
+  );
+  const missing = params.teamIds.filter((id) => !validIds.has(id));
+  if (missing.length > 0) {
+    throw new ApiError(400, `Unknown team id(s): ${missing.join(", ")}`);
+  }
+}
+
+/**
+ * Run a skill write, converting a `skill_team` foreign-key violation — a team
+ * deleted between {@link assertSkillTeams} and the insert — into a clean 400.
+ */
+async function withTeamFkErrorMapped<T>(
+  operation: () => Promise<T>,
+): Promise<T> {
+  try {
+    return await operation();
+  } catch (error) {
+    if (isForeignKeyConstraintError(error)) {
+      throw new ApiError(
+        400,
+        "One or more of the selected teams no longer exist",
+      );
+    }
+    throw error;
+  }
+}
+
+/**
+ * Authorize creating/moving a skill to the given scope and teams. Enforces the
+ * 3-tier scope check and, for non-admins, that every assigned team is one the
+ * user belongs to.
+ */
+function authorizeSkillScope(params: {
+  checker: SkillPermissionChecker;
+  scope: ResourceVisibilityScope;
+  authorId: string | null;
+  requestedTeamIds: string[];
+  userTeamIds: string[];
+  userId: string;
+}): void {
+  requireSkillModifyPermission({
+    checker: params.checker,
+    scope: params.scope,
+    authorId: params.authorId,
+    skillTeamIds: params.requestedTeamIds,
+    userTeamIds: params.userTeamIds,
+    userId: params.userId,
+  });
+
+  if (!params.checker.isAdmin && params.scope === "team") {
+    const userTeamIdSet = new Set(params.userTeamIds);
+    if (params.requestedTeamIds.some((id) => !userTeamIdSet.has(id))) {
+      throw new ApiError(
+        403,
+        "You can only assign skills to teams you are a member of",
+      );
+    }
+  }
 }
 
 function parseManifestOrThrow(raw: string) {
