+++ platform/shared/archestra-mcp-server.ts
@@ -114,7 +114,6 @@ export const TOOL_ACTIVATE_SKILL_SHORT_NAME = "activate_skill";
 export const TOOL_READ_SKILL_FILE_SHORT_NAME = "read_skill_file";
 export const TOOL_CREATE_SKILL_SHORT_NAME = "create_skill";
 export const TOOL_UPDATE_SKILL_SHORT_NAME = "update_skill";
-export const TOOL_DRAFT_SKILL_FROM_AGENT_SHORT_NAME = "draft_skill_from_agent";
 export const TOOL_CREATE_SKILL_SANDBOX_SHORT_NAME = "create_skill_sandbox";
 export const TOOL_RUN_SKILL_COMMAND_SHORT_NAME = "run_skill_command";
 export const TOOL_GET_SKILL_SANDBOX_ARTIFACT_SHORT_NAME =
@@ -190,7 +189,6 @@ export const ARCHESTRA_TOOL_SHORT_NAMES = [
   TOOL_READ_SKILL_FILE_SHORT_NAME,
   TOOL_CREATE_SKILL_SHORT_NAME,
   TOOL_UPDATE_SKILL_SHORT_NAME,
-  TOOL_DRAFT_SKILL_FROM_AGENT_SHORT_NAME,
   TOOL_CREATE_SKILL_SANDBOX_SHORT_NAME,
   TOOL_RUN_SKILL_COMMAND_SHORT_NAME,
   TOOL_GET_SKILL_SANDBOX_ARTIFACT_SHORT_NAME,
@@ -345,8 +343,6 @@ export const TOOL_CREATE_SKILL_FULL_NAME =
   `${ARCHESTRA_TOOL_PREFIX}${TOOL_CREATE_SKILL_SHORT_NAME}` as const;
 export const TOOL_UPDATE_SKILL_FULL_NAME =
   `${ARCHESTRA_TOOL_PREFIX}${TOOL_UPDATE_SKILL_SHORT_NAME}` as const;
-export const TOOL_DRAFT_SKILL_FROM_AGENT_FULL_NAME =
-  `${ARCHESTRA_TOOL_PREFIX}${TOOL_DRAFT_SKILL_FROM_AGENT_SHORT_NAME}` as const;
 export const TOOL_CREATE_SKILL_SANDBOX_FULL_NAME =
   `${ARCHESTRA_TOOL_PREFIX}${TOOL_CREATE_SKILL_SANDBOX_SHORT_NAME}` as const;
 export const TOOL_RUN_SKILL_COMMAND_FULL_NAME =
@@ -377,7 +373,6 @@ export const SKILL_ARCHESTRA_TOOL_SHORT_NAMES = [
   TOOL_READ_SKILL_FILE_SHORT_NAME,
   TOOL_CREATE_SKILL_SHORT_NAME,
   TOOL_UPDATE_SKILL_SHORT_NAME,
-  TOOL_DRAFT_SKILL_FROM_AGENT_SHORT_NAME,
 ] as const satisfies readonly ArchestraToolShortName[];
 
 export function isArchestraMcpServerTool(
